refactored DeadlineTracker to Application folder
