+++ backend/email-microservice/src/main/java/edu/oswego/cs/resources/EmailResources.java
@@ -1,5 +1,6 @@
 package edu.oswego.cs.resources;
 
+import edu.oswego.cs.application.DeadlineTracker;
 import edu.oswego.cs.database.AssignmentInterface;
 import edu.oswego.cs.database.CourseInterface;
 import edu.oswego.cs.services.EmailService;
